Add feature flag to compile with f32 or f64 support

+++ src/distribution/density_gradient.rs
@@ -8,14 +8,15 @@ use ndarray::{Array, ArrayView, Axis, Ix3, Ix4, Zip};
 use ndarray_parallel::prelude::*;
 use num_complex::Complex;
 use std::sync::Arc;
+use crate::Float;
 
 pub struct DensityGradient {
     fft_plan_forward: Arc<FFTPlan>,
     fft_plan_backward: Arc<FFTPlan>,
-    k_mesh: Array<Complex<f64>, Ix4>,
-    gradient: Array<Complex<f64>, Ix4>,
+    k_mesh: Array<Complex<Float>, Ix4>,
+    gradient: Array<Complex<Float>, Ix4>,
     grid_width: GridWidth,
-    density: Array<Complex<f64>, Ix3>,
+    density: Array<Complex<Float>, Ix3>,
 }
 
 impl DensityGradient {
@@ -24,7 +25,7 @@ impl DensityGradient {
 
         let mesh = get_k_mesh(grid_size, box_size);
 
-        let mut dummy: Array<Complex<f64>, Ix3> =
+        let mut dummy: Array<Complex<Float>, Ix3> =
             Array::default([grid_size.x, grid_size.y, grid_size.z]);
         let plan_forward = FFTPlan::new_c2c_inplace_3d(
             &mut dummy.view_mut(),
@@ -81,7 +82,7 @@ impl DensityGradient {
         let density = density.into_shape([n]).unwrap();
 
         // FFT normalization
-        let norm = n as f64;
+        let norm = n as Float;
         let norm = Complex::new(0., 1.) / norm;
 
         Zip::from(g.axis_iter_mut(Axis(1)))
@@ -101,12 +102,12 @@ impl DensityGradient {
             .for_each(|mut v| fft.reexecute3d(&mut v));
     }
 
-    pub fn get_gradient(&mut self, dist: &Distribution) -> ArrayView<Complex<f64>, Ix4> {
+    pub fn get_gradient(&mut self, dist: &Distribution) -> ArrayView<Complex<Float>, Ix4> {
         self.update_gradient(dist);
         self.gradient.view()
     }
 
-    pub fn get_real_gradient(&self) -> Array<f64, Ix4> {
+    pub fn get_real_gradient(&self) -> Array<Float, Ix4> {
         self.gradient.map(|v| v.re)
     }
 }
