+++ src/test_helper.rs
@@ -1,8 +1,15 @@
 #[cfg(test)]
-use std::f64::{EPSILON, MAX};
+#[cfg(feature = "single")]
+use std::f32::{EPSILON, MAX};
 
 #[cfg(test)]
-pub fn equal_floats(a: f64, b: f64) -> bool {
+#[cfg(not(feature = "single"))]
+use std::f64::{EPSILON, MAX};
+
+use crate::Float;
+
+#[cfg(test)]
+pub fn equal_floats(a: Float, b: Float) -> bool {
     if a == 0. && b == 0. {
         return true;
     }
@@ -17,7 +24,7 @@ pub fn equal_floats(a: f64, b: f64) -> bool {
 }
 
 #[cfg(test)]
-pub fn equal_floats_eps(a: f64, b: f64, eps: f64) -> bool {
+pub fn equal_floats_eps(a: Float, b: Float, eps: Float) -> bool {
     if a == 0. && b == 0. {
         return true;
     }
